Code refactored substantiallly.
Dictionaries for hashtags and dates have been eliminated, and replaced with a pandas series and a pandas dataframe, respectively. As such, speed and efficiency has been greatly improved.
Code has been moved around and been refactored to be more self-contained, as well. This was in preparation for having a jupyter notebook.